Move formatters to a separate module
