+++ python/lsst/meas/transiNet/modelPackages/__init__.py
@@ -21,4 +21,4 @@
 
 from .nnModelPackage import *
 from .utils import *
-
+from .formatters import *
