+++ tests/test_RBTransiNetInterface.py
@@ -23,7 +23,7 @@ import unittest
 
 import numpy as np
 
-from lsst.meas.transiNet import TransiNetTask
+from lsst.meas.transiNet import RBTransiNetTask
 from lsst.meas.transiNet import RBTransiNetInterface, CutoutInputs
 
 
@@ -31,11 +31,11 @@ class TestInference(unittest.TestCase):
     def setUp(self):
 
         # Create a mock TransiNetTask.
-        config = TransiNetTask.ConfigClass()
+        config = RBTransiNetTask.ConfigClass()
         config.modelPackageName = "dummy"
         config.modelPackageStorageMode = "local"
-        self.task = TransiNetTask(config=config)
         self.task.butler_loaded_weights = None  # Because in normal cases it's set in run().
+        self.task = RBTransiNetTask(config=config)
         self.interface = RBTransiNetInterface(self.task)
 
     def test_infer_single_empty(self):
